downloads: keeping english statement as is (for ease of quick updating to the latest one).
I'm wondering if I should redirect this page to the original one,
but it makes hard to go back-and-forth between download page and other
documentation pages in japanese translation. So I decided not to.
That is, I should provide a "japanese" version of the downloads page.

To make the quickest update for the latest version download, however,
the page is kept in almost English state so that quick copy&paste should
make it update properly. Fortunately the page should be almost self
explaining w/o Japanese translation.